Fix cyclic refs in the encounter form processor
