+++ src/utils/form-helper.test.ts
@@ -2,7 +2,6 @@ import {
   findConceptByReference,
   evaluateConditionalAnswered,
   evaluateFieldReadonlyProp,
-  parseToLocalDateTime,
   evaluateDisabled,
 } from './form-helper';
 import { DefaultValueValidator } from '../validators/default-value-validator';
@@ -381,35 +380,6 @@ describe('Form Engine Helper', () => {
       evaluateFieldReadonlyProp(field, false, false, true);
 
       expect(field.readonly).toBe(true);
-    });
-  });
-
-  describe('parseToLocalDateTime', () => {
-    it('should parse valid date string with time correctly', () => {
-      const dateString = '2023-06-27T14:30:00';
-      const expectedDate = new Date(2023, 5, 27, 14, 30, 0);
-      const parsedDate = parseToLocalDateTime(dateString);
-
-      expect(parsedDate).toEqual(expectedDate);
-    });
-
-    it('should parse valid date string without time correctly', () => {
-      const dateString = '2023-06-27';
-      const expectedDate = new Date(2023, 5, 27);
-      const parsedDate = parseToLocalDateTime(dateString);
-
-      // Set hours, minutes, and seconds to 0 since the input doesn't contain time
-      expectedDate.setHours(0, 0, 0, 0);
-
-      expect(parsedDate).toEqual(expectedDate);
-    });
-
-    it('should handle invalid date string format gracefully', () => {
-      const dateString = 'invalid-date-string';
-      const parsedDate = parseToLocalDateTime(dateString);
-
-      // Check if the parsedDate is an Invalid Date
-      expect(isNaN(parsedDate.getTime())).toBe(true);
     });
   });
